Restructure initial lesson content
- Reorder existing content and create a third lesson
- Adjust lesson 1 to be text only
- TODO: Validation is broken in lessons 2 and 3 with error:
Uncaught (in promise) TypeError: IPFS.createNode is not a function
    at eval (Lesson.vue?99a6:351)
