+++ src/lessons/File-Basics/01.md
@@ -1,7 +1,18 @@
-In our previous tutorials, we worked with the IPFS DAG API, which can be used with a variety of types of data. For the specific use case of file storage and sharing, we can save a lot of time by using the IPFS File API instead. It's custom-built to deal with all the complexities introduced by file systems.
 
-Before we start learning the File API, let's take a look at a new feature in the exercise box below that we didn't need in our previous tutorials. As a matter of security, web browsers don't let us change files that live on your computer. Therefore you'll need to upload one or more files to the browser that you can use throughout this tutorial.
+## Data added to IPFS is stored in a local node
 
-Within each exercise, you'll see that you can upload from your computer either by dragging and dropping or selecting from your file explorer. If you look closely at the `run` function in the code editor, you'll notice that it now takes an argument `files`. When you upload files from your computer, we'll make sure they're passed into the function as the `files` array. As long as you don't refresh your browser, these files will remain accessible for the next lesson in the tutorial, but you'll also have the option to upload different files to work with for each lesson.
+One of the most common questions for those new to IPFS is where content lives when it's added to the network.
 
-To practice, let's upload one or more files from your computer and create a console log to see what's been received by the browser as the `files` array. 
+As a peer-to-peer data storage system, IPFS allows each user (peer) to host whatever data they'd like locally. Typically you'd install IPFS on your own computer and create a new instance of IPFS (also known as a node) there. That's where your data would live locally, referenced by content addresses (CIDs). Data stored in IPFS can take many forms, but one of the most common use cases is the sharing of tradional files, which we'll learn more about in this tutorial.
+
+You could choose to share your data or files with your peers when you had a network connection, but if you were the only one hosting a particular resource, it would become unavailable to your peers when your machine went offline. Having multiple peers hosting the same files is what makes them more readily available, and using CIDs is what makes this system secure. We'll talk more about sharing in future tutorials, but for now we'll focus on how to work with files within your own IPFS instance.
+
+## Working with your IPFS node in ProtoSchool
+
+Here in our ProtoSchool tutorials, rather than using your machine itself, we're creating a new IPFS node for you in the browser each time you hit the "Submit" button in a lesson.
+
+Whenever you see `ipfs.someMethod()` in our lessons, `ipfs` is a variable that refers to your IPFS instance. When you replicate this process on your own machine, you can choose to call your node `ipfs` or `node` or `myNode` or `zebra`... the choice is yours! What's important to know is that the methods we're showing you will only affect your own instance of IPFS.
+
+We're creating your IPFS node behind the scenes so you can focus on the content of our lessons, but eventually you'll need to learn to host your own node locally. When you're ready to experiment, you can find instructions for [installing IPFS](https://docs.ipfs.io/introduction/install/) and [initializing your node](https://docs.ipfs.io/introduction/usage/) in our docs.
+
+Let's start exploring how we can work with files in IPFS!
