+++ src/lessons/File-Basics/02-exercise.md
@@ -1,3 +1,1 @@
-Let's add some files to IPFS. Since `files` is an array representing the files currently available to us in the browser, we'll need to loop through the array and `add()` each each file we find there to IPFS, adding the response we get to a new `addedFiles` array. (Only have a single file uploaded? No problem!) Once we're done, we'll return the `addedFiles` array to check that everything worked correctly.
-
-<em>Note: If you uploaded files in the previous lesson, you should see them listed within the upload box below. If you prefer to work with different files, just click "Start Over" above the upload box and select something new to populate our `files` array.</em>
+First, upload one or more files by dropping it below or clicking to make a selection from your file explorer.  Next, in the code editor, remove the comment markers that precede `return files.length` within the `run` function, which will make the function calculate how many files have been uploaded. (The "length" of an array is the number of items it contains.) Open your inspector and select `console` before you hit the "Submit" button. What do you see?
