+++ src/lessons/File-Basics/02.md
@@ -1,40 +1,5 @@
-Now that we have some files to work with, let's learn how to add them to IPFS!
+As a matter of security, web browsers don't let us change files that live on your computer. Therefore you'll need to upload one or more files to the browser that you can use throughout this tutorial.
 
-First things first. <strong>What does it mean to add a file to IPFS?</strong>
+Within each exercise, you'll see that you can upload from your computer either by dragging and dropping or selecting from your file explorer. If you look closely at the `run` function in the code editor, you'll notice that it now takes an argument `files`. When you upload files from your computer, we'll make sure they're passed into the function as the `files` array. As long as you don't refresh your browser, these files will remain accessible for the next lesson in the tutorial, but you'll also have the option to upload different files to work with for each lesson.
 
-As a peer-to-peer data storage system, IPFS allows each user (peer) to host whatever files they'd like locally. Typically you'd install the IPFS -- system? what word goes here?--  on your own computer and create a new instance of IPFS (also known as a node) there. That's where your files would live locally, referenced by their content addresses (CIDs). You could choose to share them with your peers when you had a network connection, but if you were the only one hosting a file, it would become unavailable to your peers when your machine went offline. Having multiple peers hosting the same files is what makes them more readily available, and using CIDs is what makes this system secure.
-
-We'll talk more about sharing in future tutorials, but for now we'll focus on how to work with files within your own IPFS instance. Here in our ProtoSchool tutorials, rather than using your machine itself, we're creating a new IPFS node for you in the browser each time you hit the "Submit" button in a lesson. This works great for learning, but eventually you'll need to learn to host files locally. -- put a link here --
-
-Whenever you see `ipfs.someMethod()` in our lessons, `ipfs` is a variable that refers to your IPFS instance. When you do this on your own, you can choose to call it `ipfs` or `node` or `myNode` or `zebra`... the choice is yours! What's important to know is that the methods we're showing you will only affect your own instance of IPFS.
-
-When we used the DAG API, remember that we added data with `ipfs.dag.add()`. Using the file API, we can use a similar command to add data with a more complex structure.
-
-To add a file to IPFS, we can use the `files.add()` method like so:
-
-```
-await ipfs.files.add(data)
-```
-The `data` passed can represent a file in the form of:
-
-* a Buffer instance
-* a Readable Stream
-* a Pull Stream
-* an array of objects, each of the form:
-  ```
-  {
-      path: '/directory/myfile.txt', // The file path you want (not the one where your file resides)
-      content: <data> // A Buffer, Readable Stream or Pull Stream with the contents of the file
-  }
-  ```
-
-Luckily for us, each file object in the `files` array in our browser is already in a format that works (Teri has no idea why or which of the formats above this is), so we can pass each file object just as it exists right now!
-
-When we add a file to IPFS using `files.add()`, the value returned for each added file should take this format:
-```
-{
-    "path": "directory", // if you didn't pass in a path, you'll see the hash here
-    "hash": "QmWXdjNC362aPDtwHPUE9o2VMqPeNeCQuTBTv1NsKtwypg",
-    "size": 67
-  }
-```
+To practice, let's upload one or more files from your computer and create a console log to see what's been received by the browser as the `files` array.
