+++ src/lessons/File-Basics/02.vue
@@ -4,7 +4,7 @@
             :validate="validate"
             :modules="modules"
             :exercise="exercise"
-            lessonTitle="Adding a file to IPFS">
+            lessonTitle="Working with files in ProtoSchool">
     </FileLesson>
   </div>
 </template>
@@ -15,43 +15,29 @@ import text from './02.md'
 import exercise from './02-exercise.md'
 
 const validate = async (result, ipfs) => {
-  const uploadedFiles = window.uploadedFiles || false
-  console.log('uploadedFiles is: ', uploadedFiles)
+  console.log('user submitted code')
   console.log('result is: ', result)
-  console.log('ipfs is: ', ipfs)
-  console.log("ipfs.files.ls('/', {long: true}) is: ", ipfs.files.ls('/', {long: true}))
-  console.log("await ipfs.files.ls('/', {long: true}) is: ", await ipfs.files.ls('/', {long: true}))
-  console.log("attempting to loop through in ipfs.files.fs:")
-  ipfs.files.ls('/', {long: true}, function (err, files) {
-    files.forEach((file) => {
-      console.log(file.name)
-    })
-  })
-  if (!result) {
-    return {'fail': 'You forgot to return a result. :('}
-  } else if (result) {
-      return {'success': 'You did something that might be right??'}
+  console.log('typeof result is: ', typeof result)
+  if (!result || typeof result === 'undefined') {
+    console.log('undefined fail message should appear')
+    return {'fail': "Looks like you forgot to return a result. Remember to remove the '//' on Line 5."}
+  } else if (typeof result === 'number') {
+    console.log('number success message should appear')
+    let fileCount = "1 file"
+    if (result > 1) {
+      fileCount = `${result} files`
+    }
+    return {'success': `You successfully uploaded ${fileCount}! ` + 'Check out your console in the inspector to see the `files` array your browser now has access to.'}
   } else {
-    return {'fail': 'Sad but useful message :('}
+    return {'fail': 'Something else is wrong.'}
   }
 }
 
 const code = `const run = async (files) => {
-  let addedFiles = []
-  for (let file of files) {
-    addedFiles.push(/* your code here*/)
-  }
-  return addedFiles
-}
-return run
-`
-
-const _solution = `const run = async (files) => {
-  let addedFiles = []
-  for (let file of files) {
-    addedFiles.push(await ipfs.files.add(file))
-  }
-  return addedFiles
+  console.log(files)
+  console.log(files.length)
+  /* remove the '//' on the line below to complete this challenge */
+  // return files.length
 }
 return run
 `
