+++ src/lessons/File-Basics/03.md
@@ -1,0 +1,32 @@
+In our previous tutorials, we worked with the IPFS DAG API, which can be used with a variety of types of data. For the specific use case of file storage and sharing, we can save a lot of time by using the IPFS File API instead. It's custom-built to deal with all the complexities introduced by file systems.
+
+When we used the DAG API, remember that we added data with `ipfs.dag.add()`. Using the file API, we can use a similar command to add data with a more complex structure.
+
+To add a file to IPFS, we can use the `files.add()` method like so:
+
+```
+await ipfs.files.add(data)
+```
+The `data` passed can represent a file in the form of:
+
+* a Buffer instance
+* a Readable Stream
+* a Pull Stream
+* an array of objects, each of the form:
+  ```
+  {
+      path: '/directory/myfile.txt', // The file path you want (not the one where your file resides)
+      content: <data> // A Buffer, Readable Stream or Pull Stream with the contents of the file
+  }
+  ```
+
+Luckily for us, each file object in the `files` array in our browser is already in a format that works <strong>(Teri has no idea why or which of the formats above this is)</strong>, so we can pass each file object just as it exists right now!
+
+When we add a file to IPFS using `files.add()`, the value returned for each added file should take this format:
+```
+{
+    "path": "directory", // if you didn't pass in a path, you'll see the hash here
+    "hash": "QmWXdjNC362aPDtwHPUE9o2VMqPeNeCQuTBTv1NsKtwypg",
+    "size": 67
+  }
+```
