+++ src/main.js
@@ -28,6 +28,7 @@ import LessonDataStructures04 from './lessons/Data-Structures/04.vue'
 import LessonDataStructures05 from './lessons/Data-Structures/05.vue'
 import LessonFileBasics01 from './lessons/File-Basics/01.vue'
 import LessonFileBasics02 from './lessons/File-Basics/02.vue'
+import LessonFileBasics03 from './lessons/File-Basics/03.vue'
 
 Vue.use(VueRouter)
 
@@ -55,6 +56,7 @@ const routes = [
   { path: '/blog/07', component: LessonBlog07 },
   { path: '/file-basics/01', component: LessonFileBasics01 },
   { path: '/file-basics/02', component: LessonFileBasics02 },
+  { path: '/file-basics/03', component: LessonFileBasics03 },
   { path: '*', name: '404' }
 ]
